Better resolution of relative paths.
We now distinguish between input and output paths, when resolving
relative paths in a *defaults* configuration file
- input paths are relative to the config file as before
- output paths are relative to the CWD
The idea being that output paths are generally meant for the user,
and so should not be buried in the package’s source directory.

Relative paths in user-specified local configuration files are always
relative to that file, as before.