Version 1.0 rewrite
This version is a complete rewrite of *ValConfig* v1, now built on top of Pydantic v2.
The logic for resolving resolving relative paths in configuration files is now both simpler and more powerful.
This version also switches to using TOML files for configuration, making use of the built-in support for TOML added in Python 3.11: this allows us to remove all the logic for dealing with the limitations of `configparser`, thus improving reliability and reducing code size by about 20%.